Split out files and renames
