+++ src/EditorFeatures/Core/EditorFeatures.csproj
@@ -130,4 +130,7 @@
     <PublicAPI Include="PublicAPI.Shipped.txt" />
     <PublicAPI Include="PublicAPI.Unshipped.txt" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="InternalUtilities\" />
+  </ItemGroup>
 </Project>
